Make skip button work.
