+++ osu.Desktop.VisualTests/Tests/TestCaseKeyCounter.cs
@@ -13,6 +13,7 @@ using OpenTK.Graphics;
 using osu.Framework.MathUtils;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Transformations;
+using osu.Game.Screens.Play;
 
 namespace osu.Desktop.VisualTests.Tests
 {
