+++ osu.Game/Screens/Play/KeyCounterKeyboard.cs
@@ -1,11 +1,11 @@
 ﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
 //Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-using OpenTK.Input;
 using osu.Framework.Graphics;
 using osu.Framework.Input;
+using OpenTK.Input;
 
-namespace osu.Game.Graphics.UserInterface
+namespace osu.Game.Screens.Play
 {
     public class KeyCounterKeyboard : KeyCounter
     {
