+++ osu.Game/osu.Game.csproj
@@ -67,6 +67,7 @@
     <Compile Include="Graphics\Backgrounds\Triangles.cs" />
     <Compile Include="Graphics\Cursor\CursorTrail.cs" />
     <Compile Include="Graphics\UserInterface\BackButton.cs" />
+    <Compile Include="Graphics\UserInterface\TwoLayerButton.cs" />
     <Compile Include="Modes\Objects\HitObjectParser.cs" />
     <Compile Include="Modes\Score.cs" />
     <Compile Include="Modes\ScoreProcesssor.cs" />
@@ -118,6 +119,7 @@
     <Compile Include="Screens\Multiplayer\MatchCreate.cs" />
     <Compile Include="Screens\Play\FailDialog.cs" />
     <Compile Include="Screens\Play\PlayerInputManager.cs" />
+    <Compile Include="Screens\Play\SkipButton.cs" />
     <Compile Include="Screens\Select\CarouselContainer.cs" />
     <Compile Include="Screens\Select\MatchSongSelect.cs" />
     <Compile Include="Screens\OsuGameMode.cs" />
@@ -145,10 +147,10 @@
     <Compile Include="Graphics\Cursor\OsuCursorContainer.cs" />
     <Compile Include="Graphics\Processing\RatioAdjust.cs" />
     <Compile Include="Graphics\TextAwesome.cs" />
-    <Compile Include="Graphics\UserInterface\KeyCounter.cs" />
-    <Compile Include="Graphics\UserInterface\KeyCounterKeyboard.cs" />
-    <Compile Include="Graphics\UserInterface\KeyCounterCollection.cs" />
-    <Compile Include="Graphics\UserInterface\KeyCounterMouse.cs" />
+    <Compile Include="Screens\Play\KeyCounter.cs" />
+    <Compile Include="Screens\Play\KeyCounterKeyboard.cs" />
+    <Compile Include="Screens\Play\KeyCounterCollection.cs" />
+    <Compile Include="Screens\Play\KeyCounterMouse.cs" />
     <Compile Include="Graphics\UserInterface\PercentageCounter.cs" />
     <Compile Include="Graphics\UserInterface\ScoreCounter.cs" />
     <Compile Include="Graphics\UserInterface\StarCounter.cs" />
